Added analyze module, moved model.py there
